requirement.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.5.0
backcall==0.2.0
black==20.8b1
certifi==2020.12.5
click==7.1.2
decorator==4.4.2
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
mypy-extensions==0.4.3
parso==0.8.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.17
psycopg2==2.8.6
ptyprocess==0.7.0
Pygments==2.8.1
PyJWT==2.3.0
python-dateutil==2.8.1
python-decouple
pytz==2021.3
pyzmq==22.0.3
regex==2020.11.13
six==1.15.0
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
